Trung

+++ db/migrate/20190924075404_create_events.rb
@@ -7,13 +7,12 @@ class CreateEvents < ActiveRecord::Migration[5.1]
       t.text       :document
       t.datetime   :startDate
       t.datetime   :endDate
-      t.integer    :duration
+      t.float      :duration
       t.string     :color
       t.text       :speaker
       t.string     :something
       t.references :user, foreign_key: true
       t.references :group, foreign_key: true
-
       t.timestamps
     end
   end
